+++ eslint.config.mjs
@@ -53,4 +53,12 @@ export default [
 			'yoda': 'error',
 		},
 	},
+	{
+		files: ['test/**/*.js'],
+		languageOptions: {
+			globals: {
+				...globals.jest,
+			},
+		},
+	},
 ];